+++ next/dweb-browser/MicroService/src/Ipc/Message/IpcStreamPaused.cs
@@ -70,7 +70,7 @@ sealed class IpcStreamPausedConverter : JsonConverter<IpcStreamPaused>
             switch (propName)
             {
                 case "type":
-                    type = (IPC_MESSAGE_TYPE)reader.GetInt16();
+                    type = (IPC_MESSAGE_TYPE)reader.GetInt32();
                     break;
                 case "stream_id":
                     stream_id = reader.GetString() ?? "";
